[services] move DctFile implementation to common place
